options

+++ src/api/controllers/post/delete-post.js
@@ -1,7 +1,9 @@
 import { Post } from "../../../models/index.js";
 
 export default async (req, res) => {
-  console.log(req.body);
+  if (req.method === "OPTIONS") {
+    return res.status(200).json({ body: "OK" });
+  }
 
   const { deletedCount } = await Post.deleteOne(req.body).catch((err) => {
     return res.status(500).json({ err });
